config/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
pyswisseph==2.8.0.post1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
